Fixed invoice list issue

+++ apim-billing-engine/src/main/java/org/wso2/apim/billing/dao/InvoiceDao.java
@@ -27,7 +27,7 @@ public interface InvoiceDao extends GenericDao<InvoiceEntity, Long> {
      * @param ID
      * @return User entity
      */
-    InvoiceEntity loadInvoiceByID(UserEntity user, int ID);
+    InvoiceEntity loadInvoiceByID(UserEntity user, long ID);
 
     List<InvoiceEntity> loadInvoices(UserEntity user);
 }
